Melhora da código, Página de Home, Api de DetalhesPedidos e Pedidos

+++ Backend/MercadoApi/Mercado.MOD/Pedido.cs
@@ -6,6 +6,6 @@
         public int? idUsuário { get; set; }
         public int? idEndereço { get; set; }
         public string? dataPedido { get; set; }
-        public int? ValorTotal { get; set; }
+        public float? ValorTotal { get; set; }
     }
 }
